+++ src/MicroweberPackages/User/database/migrations/2023_11_03_000001_add_two_factor_secret_to_users_table.php
@@ -1,0 +1,31 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddTwoFactorSecretToUsersTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        try {
+            Schema::table('users', function (Blueprint $table) {
+                if (!Schema::hasColumn('users', 'two_factor_secret')) {
+                    $table->text('two_factor_secret')->nullable();
+                }
+                if (!Schema::hasColumn('users', 'two_factor_recovery_codes')) {
+                    $table->text('two_factor_recovery_codes')->nullable();
+                }
+            });
+        } catch (Exception $e) {
+
+        }
+    }
+
+
+}