+++ userfiles/modules/site_stats/src/Models/StatsEvent.php
@@ -1,0 +1,10 @@
+<?php
+namespace MicroweberPackages\SiteStats\Models;
+
+
+use Illuminate\Database\Eloquent\Model;
+
+class StatsEvent extends Model
+{
+    protected $table = 'stats_events';
+}